Inherit sorted recent at top from application record
