+++ app/models/port.rb
@@ -1,7 +1,3 @@
 class Port < ApplicationRecord
   has_many :ships
-
-  def self.sorted_recent_at_top
-    return order(created_at: :desc)
-  end
 end
